% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cox_data}
\alias{cox_data}
\title{cox_data Univariate cox proportional hazards regression data.}
\format{
An object of class \code{data.frame} with 20 rows and 293 columns.
}
\usage{
cox_data
}
\description{
A data frame for implementing univariate cox proportional hazards regression.
}
\keyword{datasets}
