% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dot_data}
\alias{dot_data}
\title{dot_data A data frame.}
\format{
An object of class \code{data.frame} with 350 rows and 4 columns.
}
\usage{
dot_data
}
\description{
The data frame was used to plot a dot plot.
}
\keyword{datasets}
