% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dotplot}
\alias{dotplot}
\title{Plotting the Dot plot.}
\usage{
dotplot(data, low_col = "#6ADD26", high_col = "#AB2513", cut_point = 5)
}
\arguments{
\item{data}{A pathway activity score matrix, which rows represent the pathways and the columns are samples.}

\item{low_col, high_col}{Colours for low and high ends of the gradient.}

\item{cut_point}{The threshold of HRs,when HR is greater than the cut_point,HR is assigned cut_point.}
}
\value{
A Dot plot
}
\description{
The function is used to draw a graph to reflect the univariate HRs and P-values of the pathways in different cancer types.
}
\examples{
#load the data
data(dot_data)
#perform the function `dotplot`.
dotplot(dot_data)
}
