% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_indicate}
\alias{get_indicate}
\title{Indicator function.}
\usage{
get_indicate(vector, cutpoint)
}
\arguments{
\item{vector}{A numerical value.}

\item{cutpoint}{The threshold of the indicator function.}
}
\value{
An integer with the value 1 or 0.
}
\description{
Indicator function.
}
