% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mountain_plot}
\alias{mountain_plot}
\title{Plotting the density ridges plot.}
\usage{
mountain_plot(data, sample_class, Path_name)
}
\arguments{
\item{data}{A pathway activity score matrix,which row names represent the pathways and the column names are samples.}

\item{sample_class}{A vector containing subtype labels of the samples.}

\item{Path_name}{The names of the pathways that you want to show in the graph.The 'Path_name' must be included in the row names of the data.}
}
\value{
Density ridges plot
}
\description{
The function `mountain_plot` is used to draw a graph to reflect the distribution of the data.
}
\examples{
#load the data
data(Path_ES,sample_class)
#perform the function `mountain_plot`.
mountain_plot(data=Path_ES,sample_class=sample_class,Path_name=rownames(Path_ES)[c(12,20,74)])
}
