% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_onco}
\alias{mut_onco}
\title{mut_onco A binary mutation matrix.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 11968 rows and 44 columns.
}
\usage{
mut_onco
}
\description{
mut_onco is used to calculate the mutation frequency of genes..
}
\keyword{datasets}
