% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_status}
\alias{mut_status}
\title{mut_status A binary mutation matrix.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12802 rows and 50 columns.
}
\usage{
mut_status
}
\description{
The mut_status is a binary mutation matrix,in which 1 represents mutation and 0 represents wild type.
}
\keyword{datasets}
