% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Path_ES}
\alias{Path_ES}
\title{Path_ES Single-sample mutation-based pathway enrichment score profiles.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 215 rows and 882 columns.
}
\usage{
Path_ES
}
\description{
The Path_ES is used for clustering and other analysis .
}
\keyword{datasets}
