% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RWR_res}
\alias{RWR_res}
\title{RWR_res Random walk results based on mutated genes in a single sample.}
\format{
An object of class \code{numeric} of length 12436.
}
\usage{
RWR_res
}
\description{
The RWR_res is random walk results,used for enrichment analysis .
}
\keyword{datasets}
