% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samp_class_onco}
\alias{samp_class_onco}
\title{samp_class_onco}
\format{
An object of class \code{integer} of length 44.
}
\usage{
samp_class_onco
}
\description{
Sample subtype labels.
}
\keyword{datasets}
