% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_class}
\alias{sample_class}
\title{sample_class The subtype labels of samples.}
\format{
An object of class \code{integer} of length 882.
}
\usage{
sample_class
}
\description{
The subtype labels of samples.
}
\keyword{datasets}
