% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seeds_Score}
\alias{Seeds_Score}
\title{Seeds_Score The local weight of seed nodes.}
\format{
An object of class \code{data.frame} with 35 rows and 2 columns.
}
\usage{
Seeds_Score
}
\description{
Seeds_Score
}
\keyword{datasets}
