% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sur_onco}
\alias{sur_onco}
\title{sur_onco}
\format{
An object of class \code{data.frame} with 44 rows and 2 columns.
}
\usage{
sur_onco
}
\description{
Patient survival-related information.
}
\keyword{datasets}
